refactor(runtime): move __tests__ to package root
- Consolidate all test files into packages/runtime/__tests__/
- Update imports to reference src/ paths
- Move fixtures to __tests__/fixtures/
